Fix deployed files (#72)
Due to some recent schema changes, the old deployment script had to be
updated as it was modifying the deployed schema structure.

The updated script will change the indentation as the script was
rewritten to Python using a library that prettifies the output, but it
doesn't affect the schema output.

---------

Co-authored-by: Actions Bot <scripts-bot@rasentsolutions.com>