Update workflow triggers for dispatch action (#286)
Add stylesheet-repository new field to dispatch action.

This new field will be sent to the Scripts repo. The actions bot will be
able to checkout the stylesheet repository and automatically update the
schema submodules.
This way, the development server can be automatically update on every
new merged and approved PR.